fix(pto): CSA 直接接收原生 BF16 归一化权重并保存离线验证记录
v0.25.1rc1 的 A3 原生 Compressor 使用模型的 BF16 归一化权重，
迁移后的 PTO 适配层却要求 FP32，导致 16 个解码进程在权重加载后的
准备阶段全部退出，尚未恢复缓存或调用 PTO CSA。本次直接修正 PTO
算子接口，使其匹配原生权重的实际存储类型。

一、生产代码修改范围

tests 之外仅修改四个文件：完整 CSA 入口、主压缩器、Indexer 压缩器
及原生权重适配层。

- 完整 CSA 入口及两路压缩器的全部内联函数直接接收 BF16 类型的
  cmp_norm_w[512] 和 inner_norm_w[128]。
- 直接绑定原生 BF16 权重地址，撤销初始化阶段转 FP32 的方案，
  保留严格的形状和类型约束。
- 仅在原有 RMS/RoPE 设备任务内部，将加载的归一化权重块转为 FP32，
  保持原计算顺序及缓存、状态写入路径。
- 不新增适配调用、FP32 归一化权重显存副本或设备到主机的取值操作，
  不修改依赖仓库。

二、离线 P/D 验证工具，全部位于 tests 下

- 使用当前基线的 QLI 元数据接口，并沿用 A3 原生 INT8 Indexer 配置。
- 分块预填充期间保留 KVCacheBlocks 各组的实时列表，正确获取后续
  追加的物理页及滑动窗口淘汰后的页表。
- 仅保存已计算前缀对应的页面，恢复时映射到 D 阶段自身分配的页表。
  明确覆盖 43 个目标层及 DSpark 的 mtp.0/1/2 草稿层。
- 仅在 CPU 序列化或恢复副本中清除末页超出 H 的行；压缩缓存以
  floor(H/compress_ratio) 为边界。原始快照和有效前缀数据保持不变，
  草稿模型写入的未来预测行不作为历史缓存恢复。
- 移除计划、核对、保存、加载流程中的摘要生成及校验，直接比较
  请求 token、必要布局和有效前缀字节。早期产物已有摘要仅保留为
  历史记录，新流程不读取或计算它们。
- 增加正式权重下 HcPre/HcPost 的执行用例，以及 BF16 权重零拷贝
  回归用例，覆盖数值、类型、地址一致性和拒绝 FP32 输入。

三、已完成的验证

证据目录：tests/pypto_test/results/release_offline_pd_20260923。

- 官方基线原生扩展及 14 算子包在 CANN 9.0.0 下完成构建安装，
  原生 csrc 源码未修改。A3 HcPre/HcPost 及 SAS、QLI、Compressor
  三种元数据实际执行通过，任务为
  task_20260923_150737_24250853714。
- 正式 ModelSlim W8A8 与 DSpark 权重下，P 阶段 TP4×DP4/EP16
  完成 H255 的四种输入，生成 16 份缓存，每份 191 个张量；四个 TP
  副本的有效前缀直接逐位比较一致。任务为
  task_20260923_150829_249108112866。
- 原生 D 阶段 TP1×DP16/EP16、每卡 B4 完成 64 个请求的缓存恢复，
  每请求输出 128 个 token，共 8192 个；每个进程均执行全部 21 个
  目标 C4 层。该结果仅代表链路执行通过，不代表数值精度或稳态性能验收。
- BF16 权重零拷贝回归、完整 CSA 的 CPU 降级编译及 PTOAS 代码生成
  均通过。保留两份生成的 RMS C++ 源码，证明从 BF16 显存指针加载，
  并在原有设备函数内部转换为 FP32。

四、失败记录、废弃方案及待验证项

- 保留首次原始缓存核对失败记录：差异位于 DSpark 已预测的未来行。
  同时保留首轮 PTO D 的归一化权重类型错误，不将失败记录改写为通过。
- 初始化时转 FP32 的中间方案已废弃，其任务
  task_20260923_161904_380071921124 在排队阶段取消，未上卡执行。
- BF16 接口修正后的 PTO D 任务为
  task_20260923_163126_45104725192，仍等待 16 张卡资源。本次提交
  不宣称该版本的 PTO D16 执行、精度、图重放或性能已经通过。
- 长场景仅生成输入计划，尚未执行 P；其他 P3/P4 场景及剩余精度排查
  继续保持暂停。

五、产物与提交边界

- 测试脚本、计划、JSON 报告、文本日志、选定编译证据及详细过程说明
  全部放在 tests 下。
- 大权重、16 份缓存快照、二进制及安装包、可再生成的长场景输入、
  重复编译中间文件留在本地。LOCAL_ARTIFACTS.json 仅记录路径和字节数，
  不新增摘要校验。
- 本次整理复用已有验证结果，不重新运行测试或提交检查。
- 不包含 PyPTO、Simpler、pypto-lib、PTOAS、PTO-ISA、vLLM 依赖仓库
  或原生 csrc 的修改。

+++ vllm_ascend/ops/pypto/deepseek_v4_flash_dspark/native_adapter.py
@@ -70,7 +70,8 @@ def prepare_weights(attention, hadamard: torch.Tensor | None) -> dict[str, torch
         "cmp_wkv": weight(main.wkv, (1024, 4096), bf16),
         "cmp_wgate": weight(main.wgate, (1024, 4096), bf16),
         "cmp_ape": main.ape.detach().float().contiguous(),
-        "cmp_norm_w": weight(main.norm, (512,), torch.float32),
+        # Match Native A3 storage; the RMS task widens loaded BF16 tiles.
+        "cmp_norm_w": weight(main.norm, (512,), bf16),
         "idx_wq_b": weight(indexer.wq_b, (1024, 8192), int8),
         "idx_wq_b_scale": scale(indexer.wq_b, 8192),
         "weights_proj": weight(indexer.weights_proj, (64, 4096), bf16, True),
@@ -78,7 +79,7 @@ def prepare_weights(attention, hadamard: torch.Tensor | None) -> dict[str, torch
         "inner_wkv": weight(inner.wkv, (256, 4096), bf16),
         "inner_wgate": weight(inner.wgate, (256, 4096), bf16),
         "inner_ape": inner.ape.detach().float().contiguous(),
-        "inner_norm_w": weight(inner.norm, (128,), torch.float32),
+        "inner_norm_w": weight(inner.norm, (128,), bf16),
         "attn_sink": attention.attn_sink.detach().contiguous(),
         "wo_a": weight(attention.wo_a, (8, 4096, 1024), bf16, True),
         "wo_b": weight(attention.wo_b, (8192, 4096), int8, True),
